+++ src/mkvtoolnix-gui/main_window/preferences_dialog.h
@@ -89,6 +89,7 @@ protected:
   void setupPlaylistScanningPolicy();
   void setupOutputFileNamePolicy();
   void setupTrackPropertiesLayout();
+  void setupEnableMuxingTracksByType();
   void setupEnableMuxingTracksByLanguage();
   void setupMergeAddingAppendingFilesPolicy();
   void setupHeaderEditorDroppedFilesPolicy();